Chore : tabs 컴포넌트 타입 및 상수 분리
